--- CNAME
@@ -1,1 +1,0 @@
-defi-haulers.io
