+++ README.markdown
@@ -1,5 +1,5 @@
-# DeFi Haulers3.0
+# DeFi Haulers2.0
 
-A Pen created on CodePen.io. Original URL: [https://codepen.io/sendicecream/pen/8560c2a90103e22a95a304998fac5172](https://codepen.io/sendicecream/pen/8560c2a90103e22a95a304998fac5172).
+A Pen created on CodePen.io. Original URL: [https://codepen.io/sendicecream/pen/6d91ada27eafc220723be4a03118d813](https://codepen.io/sendicecream/pen/6d91ada27eafc220723be4a03118d813).
