+++ dh-frontend/src/index.js
@@ -1,0 +1,329 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './style.css';
+import MintButton from './components/MintButton'
+import ConnectButton from './components/ConnectButton'
+import { Mainnet, DAppProvider, useEtherBalance, useEthers, Config } from '@usedapp/core'
+
+ReactDOM.render(
+<React.StrictMode>
+    <DAppProvider>
+        <App>
+        </App>
+    </DAppProvider>
+</React.StrictMode>
+,
+  document.getElementById('root')
+);
+
+export function App() {
+  const { activateBrowserWallet, account } = useEthers()
+  const etherBalance = useEtherBalance(account)
+
+  return (
+    <div>
+      <link href="https://fonts.googleapis.com" rel="preconnect" />
+          <link crossOrigin="" href="https://fonts.gstatic.com" rel="preconnect" />
+              <link href="https://fonts.googleapis.com/css2?family=Caveat:wght@700&family=Ropa+Sans:ital@0;1&display=swap" rel="stylesheet" />
+                  <header>
+                      <nav className="navbar fixed-top navbar-toggleable-sm navbar-inverse bg-inverse">
+                          <div className="container">
+                              <button aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation" className="navbar-toggler navbar-toggler-right" data-target="#navbarNav" data-toggle="collapse" type="button">
+                                  <span className="navbar-toggler-icon">
+                                  </span>
+                              </button>
+                              <a className="navbar-brand" href="#">
+                                  <img src="https://assets.codepen.io/7282469/Asset+3_1.svg">
+                                  </img>
+                              </a>
+                              <div className="collapse navbar-collapse" id="navbarNav">
+                                  <ul className="navbar-nav">
+                                      <li className="nav-item active">
+                                          <a className="nav-link scroll" href="#roadmap">
+                                              Roadmap
+                                          </a>
+                                      </li>
+                                      <li className="nav-item">
+                                          <a className="nav-link scroll" href="#inventory">
+                                              Inventory
+                                          </a>
+                                      </li>
+                                      <li className="nav-item">
+                                          <a className="nav-link scroll" href="#pricing">
+                                              Trade-In
+                                          </a>
+                                      </li>
+                                      <li className="nav-item">
+                                          <a className="nav-link scroll" href="#roadmap">
+                                              About
+                                          </a>
+                                      </li>
+                                      <li className="nav-item">
+                                          <a className="nav-link scroll" href="#Twitter" style={{"color": "#1DA1F2"}}>
+                                              Twitter
+                                          </a>
+                                      </li>
+                                      <li className="nav-item">
+                                          <a className="nav-link scroll" href="#Discord" style={{"color": "#7289DA"}}>
+                                              Discord
+                                          </a>
+                                      </li>
+                                  </ul>
+                              </div>
+                              <ConnectButton />
+                          </div>
+                      </nav>
+                  </header>
+                  <section className="clickbait" id="pricing">
+                      <div className="wrapper">
+                          <section className="columns">
+                              <div className="column">
+                                <br />
+                                  <br />
+                                    <br />
+                                      <br />
+                                        <center>
+                                            <MintButton />
+                                            <br />
+                                              <br />
+                                              
+                                            
+                                        </center>
+                                      
+                                    
+                                  
+                                
+                              </div>
+                          </section>
+                      </div>
+                  </section>
+                  <section className="clickbait2" id="roadmap">
+                      <div className="wrapper">
+                          <section className="columns">
+                              <div className="column">
+                                  <h2>
+                                      Hauler
+                [ haw-ler ]
+                                  </h2>
+                                  <h6>
+                                      noun
+                                  </h6>
+                                  <h6>
+                                      :
+                                  </h6>
+                                  <br />
+                                      <div className="wrapper">
+                                          <span className="h7">
+                                              <span className="h8">
+                                                  1.
+                                              </span>
+                                              A person who hauls.
+                                              <br />
+                                                  <span className="h8">
+                                                      2.
+                                                  </span>
+                                                  A commercial trucking company.
+                                                  <br />
+                                                      <span className="h8">
+                                                          3.
+                                                      </span>
+                                                      A vehicle used for hauling or trucking.
+                                                      <br />
+                                                          <span className="h8">
+                                                              4.
+                                                          </span>
+                                                          Slang. A car capable of very high speeds.
+                                                  
+                                              
+                                          </span>
+                                      </div>
+                                  
+                              </div>
+                          </section>
+                      </div>
+                  </section>
+          <br />
+          <br />
+      <div className="wrapper">
+          <header>
+              <h3>
+                  Roadmap
+              </h3>
+          </header>
+          <section className="columns">
+              <div className="column">
+                  <h2>
+                      Launch + Mint Open
+                  </h2>
+                  <h1>
+                      DeFi Haulers launched + opened minting on January 25th, 2022.
+                  </h1>
+                  <h1>
+                      1 AVAX per mint. No max per wallet. A total of 6,210 Haulers are ready be minted into existence. Full rarity table, metadata, and creation process information will be released on mint close.
+                  </h1>
+              </div>
+              <div className="column">
+                  <h2>
+                      haulers.online
+                  </h2>
+                  <h1>
+                      DeFi Haulers are intended to serve as both collectible NFT’s and playable in-game vehicles for haulers.online - a multiplayer “top down” style, driving based game world where players battle against each other in various mini-games to earn AVAX rewards. More information and a thorough roadmap will be released for haulers.online when DeFi Haulers minting has concluded. The project is still early in development and currently a one dev show. Join the Discord or follow on Twitter for progress on game development + DeFi Haulers announcements.
+                  </h1>
+              </div>
+          </section>
+          <header>
+              <h3>
+                  Rewards Distribution
+              </h3>
+          </header>
+          <section className="columns">
+              <div className="column">
+                  <h2>
+                      40% - Trade-In Program
+                  </h2>
+                  <h1>
+                      When all 6,210 Haulers have been minted, Degen Used Motors - our automated appraiser, will begin accepting all trade-in's of pre-owned Haulers. This gives anyone the option to trade any Hauler in exchange for 0.5 AVAX - no questions asked. Trade-in vehicles are sent to the scrap yard (burned).
+                  </h1>
+              </div>
+              <div className="column">
+                  <h2>
+                      30% - Dev Wallet
+                  </h2>
+                  <h1>
+                      The dev allocation of mint sales will be used to fund haulers.online game development - in addition to contiued building + support for DeFi Haulers as a community and a brand full-time. These funds will also allow on-boarding of additional dev / community relations talent.
+                  </h1>
+              </div>
+          </section>
+          <section className="columns">
+              <div className="column">
+                  <h2>
+                      10% - Twitter Giveaway
+                  </h2>
+                  <h1>
+                      We are giving away 621 AVAX to 621 @DeFiHaulers Twitter followers. To qualify, you must follow @DeFiHaulers and post one tweet with the hastag #DeFiHaulers - and that's it. one tweet.
+                  </h1>
+              </div>
+              <div className="column">
+                  <h2>
+                      10% - Discord Giveaway
+                  </h2>
+                  <h1>
+                      The first 621 people who join and accept a role in the
+                      <br />
+                          DeFi Haulers Discord will receive 1 AVAX.
+                      
+                  </h1>
+              </div>
+          </section>
+      </div>
+      <section className="clickbait3" id="pricing">
+          <div className="wrapper">
+              <section className="columns">
+                  <div className="card">
+                      <div className="card-avatar">
+                          <img src="https://assets.codepen.io/7282469/w1.png?width=105&format=auto" />
+                              <div className="card-details">
+                                  Issued:
+                                  <br />
+                                      Exp: 02/02/2022
+                                      <br />
+                                          Please sign below:
+                              </div>
+                              <span className="label">
+                                  Upgrayedd
+                              </span>
+                      </div>
+                      <div className="card-details">
+                          <div className="name">
+                              <h6>
+                                  DeFi Haulers
+                              </h6>
+                              <h4>
+                                  Driver Identification
+                              </h4>
+                          </div>
+                          <div className="card-about">
+                              <div className="item">
+                                  <span className="label">
+                                      Alias:
+                                  </span>
+                                  <span className="value">
+                                      "dev"
+                                  </span>
+                              </div>
+                              <div className="item">
+                                  <span className="label">
+                                      Sex:
+                                  </span>
+                                  <span className="value">
+                                      M
+                                  </span>
+                              </div>
+                              <div className="item">
+                                  <span className="label">
+                                      Location:
+                                  </span>
+                                  <span className="value">
+                                      WI, USA
+                                  </span>
+                              </div>
+                          </div>
+                          <div className="skills">
+                              <span className="label">
+                                  Telegram:
+                              </span>
+                              <span className="value">
+                                  @sendicecream
+                              </span>
+                          </div>
+                          <div className="skills">
+                              <span className="label">
+                                  Instagram:
+                              </span>
+                              <span className="value">
+                                  @jeffreybuchanan
+                              </span>
+                          </div>
+                      </div>
+                  </div>
+                  <section className="clickbait2" id="pricing">
+                      <div className="column1">
+                          <h2>
+                              Welcome Drivers,
+                          </h2>
+                          - my name is
+                          <h6>
+                              Upgrayedd
+                          </h6>
+                          .
+                          <br />
+                              Strong background in design + product development.
+                              <br />
+                                  Repeat rugpull victim. Recovering shit-coin addict.
+                                  <br />
+                                      I hope you like the pictures.💜
+                                  
+                              
+                          
+                      </div>
+                  </section>
+              </section>
+          </div>
+      </section>
+      <section className="clickbait4" id="pricing">
+          <div className="wrapper">
+              <section className="columns">
+              </section>
+          </div>
+      </section>
+      <div className="footer">
+          <h1>
+              © DeFi Haulers LLC | 2022
+          </h1>
+      </div>
+      <script data-cf-beacon='{"token": "871c0057683e4ee0a83d41a8635bdf57"}' defer="" src="https://static.cloudflareinsights.com/beacon.min.js">
+      </script>
+      <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.6/js/bootstrap.min.js"></script>
+</div>
+)
+}
