--- index.template.html
@@ -1,276 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" >
-<head>
-  <meta charset="UTF-8">
-  <title>CodePen - DeFi Haulers3.0</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css">
-<link rel='stylesheet' href='https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.6/css/bootstrap.min.css'><link rel="stylesheet" href="./style.css">
-
-</head>
-<body>
-<!-- partial:index.partial.html -->
-<link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Caveat:wght@700&family=Ropa+Sans:ital@0;1&display=swap" rel="stylesheet">
-
-<!-- /start nav -->
-
-<header>
-
-	<nav class="navbar fixed-top navbar-toggleable-sm navbar-inverse bg-inverse">
-		<div class="container">
-
-			<button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-				<span class="navbar-toggler-icon"></span>
-			</button>
-
-			<a class="navbar-brand" href="#"> <img src="https://assets.codepen.io/7282469/Asset+3_1.svg"></a>
-			<div class="collapse navbar-collapse" id="navbarNav">
-
-				<ul class="navbar-nav">
-
-					<li class="nav-item active">
-						<a class="nav-link scroll" href="#roadmap">Roadmap</a>
-					</li>
-
-					<li class="nav-item">
-						<a class="nav-link scroll" href="#inventory">Inventory</a>
-					</li>
-					<li class="nav-item">
-						<a class="nav-link scroll" href="#pricing">Trade-In</a>
-					</li>
-					<li class="nav-item">
-						<a class="nav-link scroll" href="#roadmap">About</a>
-					</li>
-					<li class="nav-item">
-						<a class="nav-link scroll" href="#Twitter" style="color:#1DA1F2">Twitter</a>
-					</li>
-					<li class="nav-item">
-						<a class="nav-link scroll" href="#Discord" style="color:#7289DA">Discord</a>
-					</li>
-
-			</div>
-			<button class="enableEthereumButton btn btn-primary">CONNECT</button>
-		</div>
-
-	</nav>
-
-</header>
-
-<!-- /end nav -->
-
-<!-- /start splash banner -->
-
-<section class="clickbait" id="pricing">
-
-	<div class="wrapper">
-
-		<div class="column">
-
-		</div>
-	</div>
-
-</section>
-
-<!-- /end splash banner -->
-
-<section class="clickbait2">
-
-	<div class="wrapper">
-
-		<div class="column">
-			<p>Hauler
-				[ haw-ler ]</p>
-			<h8>noun</h8>
-			<h7>:</h7>
-			<br>
-			<div class="wrapper">
-
-				<h6>
-
-					<h8>1. </h8> A person who hauls.
-
-					<br>
-
-					<h8>2. </h8> A commercial trucking company.
-
-					<br>
-
-					<h8>3. </h8> A vehicle used for hauling or trucking.
-
-					<br>
-
-					<h8>4. </h8> Slang. A car capable of very high speeds.
-
-				</h6>
-			</div>
-</section>
-
-</div>
-</section>
-
-<!-- /end about -->
-
-<!-- /start roadmap -->
-<section class="clickbait0" id="Roadmap">
-
-	<div class="wrapper">
-
-		<header>
-			<h3>Roadmap </h3>
-		</header>
-
-		<section class="columns">
-
-			<div class="column">
-				<h3>1. </h3>
-				<h2>Launch + Mint Open</h2>
-				<h1>DeFi Haulers launched + opened mint on January 28th, 2022. </h1>
-
-				<h1> 1 AVAX per mint. No max per wallet. A total of 6,210 Haulers are ready to be minted into existence. </h1>
-	
-					<br>
-
-					<h10>
-	Full rarity breakdown, metadata, and creation process information will be released on mint close. </h10>
-			</div>
-
-			<div class="column">
-				<h3>2. </h3>
-				<h2>Degen Motors</h2>
-
-				<h1> When all 6,210 Haulers have been dispatched - Degen Motors will begin accepting trade-in's. Hauler owners may choose to trade any Hauler at any time in exchange for 0.5 AVAX. This is an automated process in which users instantly receive the 0.5 AVAX payout. Trade-In Haulers are sent directly to the scrap yard (burned). </h1>
-			</div>
-
-			<div class="column">
-				<h3>3. </h3>
-				<h2>haulers.online</h2>
-				<h1>DeFi Haulers exist to serve as both collectible NFT’s and playable in-game vehicles in haulers.online - a browser based, multi-player, “top down” style, driving based game world. The game is four months into prototyping inside Unreal Engine 5 - with primary focus on core game mechanics + art direction. More details TBA.
-
-			</div>
-
-		</section>
-
-</section>
-</div>
-
-<!-- /end roadmap -->
-
-<!-- /start inventory -->
-
-<section class="clickbait5" id="inventory">
-
-	<div class="wrapper">
-
-		<header>
-			<h3>Inventory</h3>
-
-		</header>
-</section>
-</section>
-</div>
-
-<!-- /end inventory -->
-
-<!-- /start about -->
-
-<section class="clickbait3" id="pricing">
-
-	<div class="wrapper">
-		<section class="columns">
-			<div class="card">
-				<div class="card-avatar">
-					<img src="https://assets.codepen.io/7282469/w1.png?width=105&format=auto">
-
-					<div class="card-details">Issued: <br> Exp: 02/02/2022<br>Please sign below:</div>
-					<span class="label">Upgrayedd</span>
-
-				</div>
-				<div class="card-details">
-
-					<div class="name">
-						<h8>DeFi Haulers</h8>
-						<h4>Driver Identification </h4>
-					</div>
-					<div class="card-about">
-						<div class="item">
-							<span class="label">Alias:</span>
-							<span class="value">dev, jeff, JR</span>
-						</div>
-						<div class="item">
-							<span class="label">Sex:</span>
-							<span class="value">M</span>
-
-						</div>
-						<div class="item">
-							<span class="label">Location:</span>
-							<span class="value">WI, USA</span>
-
-						</div>
-
-					</div>
-					<div class="skills">
-						<span class="label">Instagram:</span>
-						<span class="value">@sendicecream</span>
-					</div>
-
-					<div class="skills">
-						<span class="label">Telegram:</span>
-						<span class="value">@sendicecream_was_taken</span>
-
-					</div>
-
-				</div>
-			</div>
-
-			<section class="clickbait2" id="pricing">
-
-				<div class="column1">
-
-					<h2>Hello Drivers, </h2><br> My name is <h8>Upgrayedd</h8>.
-					<br>
-					Strong background in design + product development.
-					<br>
-					Repeat rugpull victim. Recovering shit-coin addict.
-					<br>
-					<br>
-					I hope you like the pictures.💜
-				</div>
-
-	</div>
-</section>
-</section>
-
-<!-- /end about -->
-
-<section class="clickbait5" id="pricing">
-	<h7> Got a lemon? </h7>
-	<br>
-	Generate the pain away today with this
-	<br> complimentary DeFi Haulers Clickbait Generator.
-	<br>
-	<br>
-
-	<p id="hypebox"></p>
-	<button class="btn btn-primary">Generate</button>
-
-</section>
-<!-- /start footer -->
-
-<div class="footer">
-
-	<h1>© DeFi Haulers LLC | 2022</h1>
-
-</div>
-
-<!-- Cloudflare Web Analytics -->
-
-<script defer src='https://static.cloudflareinsights.com/beacon.min.js' data-cf-beacon='{"token": "871c0057683e4ee0a83d41a8635bdf57"}'></script>
-
-<!-- End Cloudflare Web Analytics -->
-<!-- partial -->
-  <script src='https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js'></script>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.6/js/bootstrap.min.js'></script><script  src="./script.js"></script>
-
-</body>
-</html>
