+++ src/index.partial.html
@@ -1,0 +1,294 @@
+
+
+<link rel="preconnect" href="https://fonts.googleapis.com">
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+<link href="https://fonts.googleapis.com/css2?family=Caveat:wght@700&family=Ropa+Sans:ital@0;1&display=swap" rel="stylesheet">
+
+<!-- /start nav -->
+
+<header>
+
+	<nav class="navbar fixed-top navbar-toggleable-sm navbar-inverse bg-inverse">
+		<div class="container">
+
+			<button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+				<span class="navbar-toggler-icon"></span>
+			</button>
+
+			<a class="navbar-brand" href="#"> <img src="https://assets.codepen.io/7282469/Asset+3_1.svg"> </a>
+			<div class="collapse navbar-collapse" id="navbarNav">
+
+				<ul class="navbar-nav">
+
+					<li class="nav-item active">
+						<a class="nav-link scroll" href="#roadmap">Roadmap</a>
+					</li>
+
+					<li class="nav-item">
+						<a class="nav-link scroll" href="#inventory">Inventory</a>
+					</li>
+					<li class="nav-item">
+						<a class="nav-link scroll" href="#pricing">Trade-In</a>
+					</li>
+					<li class="nav-item">
+						<a class="nav-link scroll" href="#roadmap">About</a>
+						</li>
+					<li class="nav-item">
+						<a class="nav-link scroll" href="#Twitter" style="color:#1DA1F2">Twitter</a>
+					</li>
+					<li class="nav-item">
+						<a class="nav-link scroll" href="#Discord" style="color:#7289DA">Discord</a>
+					</li>
+
+			</div>
+			<button class="enableEthereumButton btn btn-primary">CONNECT</button>
+		</div>
+
+	</nav>
+
+</header>
+
+<!-- /end nav -->
+
+<!-- /start splash banner -->
+
+	
+	
+	
+<!-- /end splash banner -->
+
+<!-- /start about -->
+
+<section class="clickbait" id="pricing">
+
+	<div class="wrapper">
+
+		<section class="columns">
+
+			<div class="column">
+
+				<br>
+				<br>
+				<br>
+				<br>
+				<center>
+					<button class="sendEthButton btn btn-success">Mint</button>
+					<br>
+					<br>
+				</center>
+
+			</div>
+	</div>
+
+</section>
+
+<section class="clickbait2" id="roadmap">
+
+	<div class="wrapper">
+
+		<section class="columns">
+
+			<div class="column">
+				<h2>Hauler
+					[ haw-ler ]</h2>	<h8>noun</h8> <h7>:</h7>
+<br>
+					<div class="wrapper">
+
+				<h6> 
+					
+					<h8>1.</h8>   A person who hauls.
+				
+
+				<br>
+					
+					
+					<h8>2.</h8>   A commercial trucking company.
+			
+				<br>
+		
+					
+					<h8>3.</h8>   A vehicle used for hauling or trucking.
+
+				
+		<br>
+					
+				<h8>4.</h8>   Slang. A car capable of very high speeds.
+				
+				</h6>
+				</div>
+				
+
+</h5>
+			
+			
+			
+			
+					<br>
+					<br>
+			</div>
+			</div>
+				</section>
+			
+			
+			
+			
+<!-- /end about -->
+
+<!-- /start roadmap -->
+
+<div class="wrapper">
+
+	<header>
+		<h3>Roadmap </h3>
+	</header>
+
+	<section class="columns">
+
+		<div class="column">
+			<h2>Launch + Mint Open</h2>
+			<h1>DeFi Haulers launched + opened minting on January 25th, 2022. </h1>
+
+			<h1> 1 AVAX per mint. No max per wallet. A total of 6,210 Haulers are ready be minted into existence. Full rarity table, metadata, and creation process information will be released on mint close.  </h1>
+		</div>
+
+		<div class="column">
+			<h2>haulers.online</h2>
+			<h1>DeFi Haulers are intended to serve as both collectible NFT’s and playable in-game vehicles for haulers.online - a multiplayer “top down” style, driving based game world where players battle against each other in various mini-games to earn AVAX rewards. More information and a thorough roadmap will be released for haulers.online when DeFi Haulers minting has concluded. The project is still early in development and currently a one dev show. Join the Discord or follow on Twitter for progress on game development + DeFi Haulers announcements. </h1>
+		</div>
+
+	</section>
+
+	<header>
+		<h3>Rewards Distribution</h3>
+	</header>
+
+	<section class="columns">
+
+		<div class="column">
+			<h2>40% - Trade-In Program</h2>
+			<h1>When all 6,210 Haulers have been minted, Degen Used Motors - our automated appraiser, will begin accepting all trade-in's of pre-owned Haulers. This gives anyone the option to trade any Hauler in exchange for 0.5 AVAX - no questions asked. Trade-in vehicles are sent to the scrap yard (burned). </h1>
+		</div>
+
+		<div class="column">
+			<h2>30% - Dev Wallet</h2>
+			<h1>The dev allocation of mint sales will be used to fund haulers.online game development - in addition to contiued building + support for DeFi Haulers as a community and a brand full-time. These funds will also allow on-boarding of additional dev / community relations talent.    </h1>
+		</div>
+
+	</section>
+
+	<section class="columns">
+
+		<div class="column">
+			<h2>10% - Twitter Giveaway</h2>
+			<h1>We are giving away 621 AVAX to 621 @DeFiHaulers Twitter followers. To qualify, you must follow @DeFiHaulers and post one tweet with the hastag #DeFiHaulers - and that's it. one tweet. </h1>
+		</div>
+
+		<div class="column">
+			<h2>10% - Discord Giveaway</h2>
+			<h1>The first 621 people who join and accept a role in the
+				<br>
+
+				DeFi Haulers Discord will receive 1 AVAX.
+			</h1>
+		</div>
+
+	</section>
+
+</div>
+
+<!-- /end rewards -->
+
+<!-- /start about -->
+
+<section class="clickbait3" id="pricing">
+
+	<div class="wrapper">
+		<section class="columns">
+
+			<div class="card">
+				<div class="card-avatar">
+					<img src="https://assets.codepen.io/7282469/w1.png?width=105&format=auto">
+
+					<div class="card-details">Issued: <br> Exp: 02/02/2022<br>Please sign below:</div>
+					<span class="label">Upgrayedd</span>
+
+				</div>
+				<div class="card-details">
+
+					<div class="name"><h8>DeFi Haulers</h8> <h4>Driver Identification </h4></div>
+					<div class="card-about">
+						<div class="item">
+							<span class="label">Alias:</span>
+							<span class="value">"dev"</span>
+</div>
+						<div class="item">
+								<span class="label">Sex:</span>
+							<span class="value">M</span>
+
+
+						</div>
+						<div class="item">
+							<span class="label">Location:</span>
+							<span class="value">WI, USA</span>
+							
+						</div>
+
+
+					</div>
+					<div class="skills">
+						<span class="label">Telegram:</span>
+						<span class="value">@sendicecream</span>
+					</div>
+
+					<div class="skills">
+						<span class="label">Instagram:</span>
+						<span class="value">@jeffreybuchanan</span>
+					</div>
+
+				</div>
+			</div>
+
+			<section class="clickbait2" id="pricing">
+
+				<div class="column1">
+
+					<h2>Welcome Drivers, </h2> - my name is <h8>Upgrayedd</h8>.
+						<br>
+						Strong background in design + product development.
+						<br>
+						Repeat rugpull victim. Recovering shit-coin addict.
+						<br>
+						I hope you like the pictures.💜
+				</div>
+ 
+	</div>
+	</div>
+</section>
+
+<!-- /end about -->
+
+<section class="clickbait4" id="pricing">
+
+	<div class="wrapper">
+		<section class="columns">
+
+		</section>
+	</div>
+</section>
+<!-- /start footer -->
+
+<div class="footer">
+
+	<h1>© DeFi Haulers LLC | 2022</h1>
+
+</div>
+	
+	
+	
+	<!-- Cloudflare Web Analytics -->
+	
+	
+	<script defer src='https://static.cloudflareinsights.com/beacon.min.js' data-cf-beacon='{"token": "871c0057683e4ee0a83d41a8635bdf57"}'></script>
+	
+	
+	<!-- End Cloudflare Web Analytics -->
